+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.0.3] - 2023-11-03
+
 - Use [`dry-validation`](https://github.com/dry-rb/dry-validation) to validate data classes
 - Use [`dry-types`](https://github.com/dry-rb/dry-types) and [`dry-struct`](https://github.com/dry-rb/dry-struct) for data classes